Separate into files & make counting more smart
Separate app.py into many files, and make that counting now accepts mathematical expressions
