+++ .gitignore
@@ -1,1 +1,2 @@
 /token
+__pycache__/
